Reorganize package names for connectors
